# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  productId: string;
  name: string;
  karat: string;
  weight: number;
  price: number;
  stock: number;
  imageUrl: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  productId: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  karat: string;
  /** @minimum 0 */
  weight: number;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stock: number;
  /** @minLength 1 */
  imageUrl: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  productId?: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  karat?: string;
  /** @minimum 0 */
  weight?: number;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  stock?: number;
  /** @minLength 1 */
  imageUrl?: string;
}

export interface OrderInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 10 */
  phone: string;
  /** @minLength 5 */
  address: string;
  /** @minimum 1 */
  productId: number;
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  dispatched: 'dispatched',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  fullName: string;
  phone: string;
  address: string;
  productId: number;
  productName: string;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  createdAt: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  dispatched: 'dispatched',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface AdminLoginInput {
  /** @minLength 10 */
  phone: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  pin: string;
}

export interface AdminSession {
  token: string;
  expiresAt: string;
}

export interface AdminSummary {
  productCount: number;
  lowStockCount: number;
  pendingOrderCount: number;
  orderValue: number;
}