# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Jewellery catalog and inventory operations
  - name: orders
    description: Customer order operations
  - name: admin
    description: Store administration operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List jewellery products
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a jewellery product
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a jewellery product
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a jewellery product
      security:
        - adminBearer: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a jewellery product
      security:
        - adminBearer: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "204":
          description: Product deleted
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List customer orders
      security:
        - adminBearer: []
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a customer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      security:
        - adminBearer: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Sign in to the store admin panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get admin dashboard summary
      security:
        - adminBearer: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, productId, name, karat, weight, price, stock, imageUrl, createdAt, updatedAt]
      properties:
        id: { type: number }
        productId: { type: string }
        name: { type: string }
        karat: { type: string }
        weight: { type: number }
        price: { type: number }
        stock: { type: number }
        imageUrl: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProductInput:
      type: object
      required: [productId, name, karat, weight, price, stock, imageUrl]
      properties:
        productId: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        karat: { type: string, minLength: 1 }
        weight: { type: number, minimum: 0 }
        price: { type: number, minimum: 0 }
        stock: { type: number, minimum: 0 }
        imageUrl: { type: string, minLength: 1 }
    ProductUpdate:
      type: object
      properties:
        productId: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        karat: { type: string, minLength: 1 }
        weight: { type: number, minimum: 0 }
        price: { type: number, minimum: 0 }
        stock: { type: number, minimum: 0 }
        imageUrl: { type: string, minLength: 1 }
    OrderInput:
      type: object
      required: [fullName, phone, address, productId, quantity]
      properties:
        fullName: { type: string, minLength: 2 }
        phone: { type: string, minLength: 10 }
        address: { type: string, minLength: 5 }
        productId: { type: number, minimum: 1 }
        quantity: { type: number, minimum: 1, maximum: 20 }
    Order:
      type: object
      required: [id, fullName, phone, address, productId, productName, quantity, totalPrice, status, createdAt]
      properties:
        id: { type: number }
        fullName: { type: string }
        phone: { type: string }
        address: { type: string }
        productId: { type: number }
        productName: { type: string }
        quantity: { type: number }
        totalPrice: { type: number }
        status: { type: string, enum: [pending, confirmed, dispatched, delivered, cancelled] }
        createdAt: { type: string }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, confirmed, dispatched, delivered, cancelled] }
    AdminLoginInput:
      type: object
      required: [phone, pin]
      properties:
        phone: { type: string, minLength: 10 }
        pin: { type: string, minLength: 6, maxLength: 6 }
    AdminSession:
      type: object
      required: [token, expiresAt]
      properties:
        token: { type: string }
        expiresAt: { type: string }
    AdminSummary:
      type: object
      required: [productCount, lowStockCount, pendingOrderCount, orderValue]
      properties:
        productCount: { type: number }
        lowStockCount: { type: number }
        pendingOrderCount: { type: number }
        orderValue: { type: number }