# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 10 */
  phone: string;
  /** @minLength 5 */
  address: string;
  /** @minimum 1 */
  productId: number;
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}