# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  productId: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  karat: string;
  /** @minimum 0 */
  weight: number;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stock: number;
  /** @minLength 1 */
  imageUrl: string;
}